[FluidAPI]remove clip api (#48946)
